tripmind-python/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
firebase-admin>=6.6.0
anthropic>=0.40.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
sse-starlette>=2.1.0
